tools: add parallel loop for validation of virtual group and preselection (#322)
This adds a parallel loop run for internal users who haven't enabled
parallel tool calling. For the first MCP tool call made per tool loop,
we try replaying that loop with virtual tools and ensure the correct
virtual tool is called to expand the group. Or, if the tool is
preselected to be expanded, then we record that, too.

This is internal-only to avoid using premium requests for paying users
on this background query.

This can be disabled via an EXP flag if we see it causes unexpected load
(but since this is internal-only I don't anticipate any issues.)